% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsicleR.R
\name{FeatureScatter_compat}
\alias{FeatureScatter_compat}
\title{FeatureScatter with Seurat version compatibility}
\usage{
FeatureScatter_compat(object, feature1, feature2, ...)
}
\arguments{
\item{object}{Seurat object}

\item{feature1}{Name of x feature (metadata column or gene)}

\item{feature2}{Name of y feature (metadata column or gene)}

\item{...}{Additional args forwarded to the plotting (currently `pt.size`, `cells`, `group.by`, `col` are supported)}
}
\value{
ggplot2 plot
}
\description{
Wrapper around Seurat's FeatureScatter to support Seurat v5 assay "layers".
If a requested feature is not found in `object@meta.data` it will try common
assay layer names ("data", "logcounts", "normalized", "counts", "scale.data").
}
\keyword{internal}
