% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsicleR.R
\name{FilterPlots}
\alias{FilterPlots}
\title{FilterPlots}
\usage{
FilterPlots(UMI, G_RNA_low = 0, G_RNA_hi = Inf, U_RNA_low = 0, U_RNA_hi = Inf, percent_mt_hi = 100, percent_ribo_hi = 100, percent_disso_hi = 100, out_folder=getwd())
}
\arguments{
\item{UMI}{Input UMI object generated by PrePlots function}

\item{G_RNA_low}{Minimum number of genes detected in each cell. Default is 0}

\item{G_RNA_hi}{Maximum number of genes detected in each cell. Default is Infinite}

\item{U_RNA_low}{Minimum number of molecules detected within a cell. Default is 0}

\item{U_RNA_hi}{Maximum number of molecules detected within a cell. Default is Infinite}

\item{percent_mt_hi}{Maximum percentage of mitochondrial genes. Default is 100}

\item{percent_ribo_hi}{Maximum percentage of ribosomal genes. Default is 100}

\item{percent_disso_hi}{Maximum percentage of dissociation genes. Default is 100}

\item{out_folder}{Output folder. Default is working directory}
}
\value{
Returns a Seurat Object filtered according to the user-defined set of thresholds.
}
\description{
Apply filters to input data. Creates plot with provided thresholds as input.
}
\details{
FilterPlots returns two graphs in the "01.QC_Plots" folder:

the first plot (01f) displays data distributions after filtering; filtering thresholds are displayed as gray lines.

the second graph (01g) shows data correlations, that highlight with colors, cells passing the filtering procedure; filtering thresholds are displayed as gray lines.
}
\examples{

FilterPlots(umi_object)

FilterPlots(umi_object, G_RNA_low = 500, G_RNA_hi = 10000, U_RNA_low = 1000, U_RNA_hi = 70000, percent_mt_hi = 20, percent_ribo_hi = 50, percent_disso_hi = 10)

}
\author{
Jimmy Caroli, Francesco Grandi
}
