% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsicleR.R
\name{Normalize}
\alias{Normalize}
\title{Normalize}
\usage{
Normalize(UMI, variable_genes=2000, out_folder=getwd())
}
\arguments{
\item{UMI}{Input UMI object generated via previous popsicleR functions}

\item{variable_genes}{Number of highly variable genes investigated after normalization. Default is 2000}

\item{out_folder}{Output folder. Default is the working directory}
}
\value{
Returns a Seurat Object after normalization process.
}
\description{
Perform data normalization on the provided input. User must define the number of variable genes to be investigated
after normalization.
}
\details{
Normalize returns two graphs in the "02.PreProcessing" folder:

the first graph (02a) shows expression before and after the normalization step,

while the second (02b) highlights the highly variable genes.
}
\examples{

Normalize(seurat_object, variable_genes=2000)

}
\author{
Jimmy Caroli, Francesco Grandi
}
