% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsicleR.R
\name{PrePlots}
\alias{PrePlots}
\title{PrePlots}
\usage{
PrePlots(sample_name, input_data, genelist=NULL, percentage=0.1, gene_filter=200, cellranger=TRUE, organism=c("human","mouse"), out_folder=getwd())
}
\arguments{
\item{sample_name}{Name of the sample investigated}

\item{input_data}{Input data for popsicleR. When data are generated via Cell Ranger input_data must be a path to a folder containing the approriate output. Otherwise, input_data must be a path to a .txt matrix structured with gene names in the first column and cell names in the first row.}

\item{genelist}{List of gene for singular gene plots. By default is NULL and a proprietary genelist is used to generate plots. A personal genelist can be fed as characters vector}

\item{percentage}{Minimum percentage of cells in which a gene must be expressed to be retained for the subsequent analysis. Default is 0.1}

\item{gene_filter}{Minimum number of genes detected to retained a cell for the subsequent analysis. Default is 200}

\item{cellranger}{Specify if data were generated via Cell Ranger. Default is TRUE}

\item{organism}{The organism on which perfom the analysis. Can be human or mouse.}

\item{out_folder}{Output folder. Default is working directory}
}
\value{
Returns a Seurat Object generated from input data.

Genes expressed in less than percentage of cells the value specified by the "percentage" parameter and cells that express less than number of genes specified by "gene_filter" parameter are filtered-out from input data.
}
\description{
Perfoms preprossing analyses and quality controls on the provided dataset.
Reports information on the dataset and outputs plots describing quality of the data to be further analyzed.
}
\details{
PrePlots returns several graphs in the "01.QC_Plots" folder:

violin (1a), density (1b), and scatter (1c) plots on number of genes (nFeature_RNA), number of UMI (nCount_RNA), the percentage of reads that map to the mitochondrial genome (percent_mt), to ribosomal (percent_ribo) and dissociation genes (percent_disso).

PrePlots also returns the distributions (1d) and the cell abundance (1e) of specific cell populations identified by a list of user-defined marker genes.
}
\examples{

PrePlots('breast_single_cell', input_data="/path/to/cellranger_output_directory/", genelist=c('TP53','PTEN'), percentage=0.1, gene_filter=200, cellranger = TRUE, organism="human")

PrePlots("mouse_sample", input_data="/path/to/matrix.txt", genelist=c('Tp53','Pten'), cellranger=FALSE, organism= "mouse")

}
\author{
Jimmy Caroli, Francesco Grandi
}
